Ticket 564: Track payload size with 64 bit integer

+++ src/network_lib/src/jaln_pub_feeder.c
@@ -40,7 +40,7 @@
 // to a different file.
 #include "jaln_subscriber_state_machine.h"
 
-axl_bool jaln_pub_feeder_get_size(jaln_session *sess, int *size)
+axl_bool jaln_pub_feeder_get_size(jaln_session *sess, uint64_t *size)
 {
 	// expect that the pub_data is already filled out...
 	*size = sess->pub_data->vortex_feeder_sz - sess->pub_data->payload_off;
@@ -221,7 +221,7 @@ void * APR_THREAD_FUNC jaln_pub_feeder_handler(
 	struct jaln_response_header_info *info = jaln_response_header_info_create(sess);
 	info->expected_nonce = jal_strdup(sess->pub_data->nonce);
 
-	int size;
+	uint64_t size;
 	jaln_pub_feeder_get_size(sess, &size);
 	curl_easy_setopt(ctx, CURLOPT_POSTFIELDSIZE_LARGE, size);
